+++ app/src/main/java/communi/dog/aplicatiion/Activities/Emergency_numbers.java
@@ -1,4 +1,4 @@
-package communi.dog.aplicatiion;
+package communi.dog.aplicatiion.Activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -7,6 +7,8 @@ import android.net.Uri;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+
+import communi.dog.aplicatiion.R;
 
 public class Emergency_numbers extends AppCompatActivity {
