+++ app/src/main/java/communi/dog/aplicatiion/Activities/MapScreenActivity.java
@@ -1,4 +1,4 @@
-package communi.dog.aplicatiion;
+package communi.dog.aplicatiion.Activities;
 
 import android.content.DialogInterface;
 import android.content.Intent;
@@ -20,6 +20,12 @@ import androidx.drawerlayout.widget.DrawerLayout;
 import com.google.android.material.navigation.NavigationView;
 
 import org.osmdroid.config.Configuration;
+
+import communi.dog.aplicatiion.CommuniDogApp;
+import communi.dog.aplicatiion.DB;
+import communi.dog.aplicatiion.MapHandler;
+import communi.dog.aplicatiion.MapState;
+import communi.dog.aplicatiion.R;
 
 
 public class MapScreenActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
