+++ app/src/main/java/communi/dog/aplicatiion/Callbacks/FirebaseMapStateUpdateCallback.java
@@ -1,4 +1,6 @@
-package communi.dog.aplicatiion;
+package communi.dog.aplicatiion.Callbacks;
+
+import communi.dog.aplicatiion.MapState;
 
 public interface FirebaseMapStateUpdateCallback {
     void onCallbackMapState(MapState mapState);
