+++ app/src/main/java/communi/dog/aplicatiion/Callbacks/FirebaseUsersUpdateCallback.java
@@ -1,6 +1,8 @@
-package communi.dog.aplicatiion;
+package communi.dog.aplicatiion.Callbacks;
 
 import java.util.HashMap;
+
+import communi.dog.aplicatiion.User;
 
 public interface FirebaseUsersUpdateCallback {
     void onCallback(HashMap<String, User> users);
