+++ app/src/main/java/communi/dog/aplicatiion/Callbacks/OnApprovalRequestCallback.java
@@ -1,0 +1,7 @@
+package communi.dog.aplicatiion.Callbacks;
+
+import communi.dog.aplicatiion.User;
+
+public interface OnApprovalRequestCallback {
+    void onClick(User user);
+}
