+++ app/src/main/java/communi/dog/aplicatiion/Callbacks/OnMapLongPressCallback.java
@@ -1,4 +1,4 @@
-package communi.dog.aplicatiion;
+package communi.dog.aplicatiion.Callbacks;
 
 import org.osmdroid.util.GeoPoint;
