--- app/src/main/java/communi/dog/aplicatiion/OnApprovalRequestClickListener.java
@@ -1,5 +1,0 @@
-package communi.dog.aplicatiion;
-
-public interface OnApprovalRequestClickListener {
-    void onClick(User user);
-}
