+++ app/src/main/res/layout/activity_emergency_numbers.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Emergency_numbers">
+    tools:context=".Activities.Emergency_numbers">
 
     <ScrollView
         android:layout_width="match_parent"
