+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:id="@+id/loginConstraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginActivity">
+    tools:context=".Activities.LoginActivity">
 
     <ImageView
         android:id="@+id/logo_image_in_login"
