+++ app/src/main/res/layout/activity_map_screen.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".MapScreenActivity"
+    tools:context=".Activities.MapScreenActivity"
     tools:openDrawer="start">
