+++ app/src/main/res/layout/activity_profile_page.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/profile_bg2"
-    tools:context=".ProfilePageActivity">
+    tools:context=".Activities.ProfilePageActivity">
 
     <!--  logo  -->
     <ImageView
