+++ app/src/main/res/layout/activity_register.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFFFFF"
-    tools:context=".RegisterActivity">
+    tools:context=".Activities.RegisterActivity">
 
     <!-- screen's logo -->
     <ImageView
