+++ app/src/main/res/layout/activity_users_approval.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UserApprovalActivity">
+    tools:context=".Activities.UserApprovalActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/unapprovedUsersRecyclerView"
